+++ src/components/popup/popup.styles.ts
@@ -5,8 +5,15 @@ export default css`
   ${componentStyles}
 
   :host {
-    --arrow-size: 4px;
     --arrow-color: var(--sl-color-neutral-1000);
+    --arrow-size: 6px;
+
+    /*
+     * These properties are computed to account for the arrow's dimensions after being rotated 45º. The constant
+     * 0.7071 is derived from sin(45), which is the diagonal size of the arrow's container after rotating.
+     */
+    --arrow-size-diagonal: calc(var(--arrow-size) * 0.7071);
+    --arrow-padding-offset: calc(var(--arrow-size-diagonal) - var(--arrow-size));
 
     display: contents;
   }
@@ -28,8 +35,8 @@ export default css`
 
   .popup__arrow {
     position: absolute;
-    width: calc(var(--arrow-size) * 2);
-    height: calc(var(--arrow-size) * 2);
+    width: calc(var(--arrow-size-diagonal) * 2);
+    height: calc(var(--arrow-size-diagonal) * 2);
     transform: rotate(45deg);
     background: var(--arrow-color);
     z-index: -1;