+++ docs/resources/changelog.md
@@ -8,6 +8,10 @@ New versions of Shoelace are released as-needed and generally occur when a criti
 
 _During the beta period, these restrictions may be relaxed in the event of a mission-critical bug._ 🐛
 
+## Next
+
+- Fixed a bug in `<sl-popup>` that didn't account for the arrow's diagonal size
+
 ## 2.0.0-beta.82
 
 - Added the `sync` and `arrow-placement` attributes to `<sl-popup>`